requirements.txt
aiogram==3.17.0
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
mistralai==1.2.5
aiofiles==24.1.0
feedparser==6.0.11
